Move migration functions out of models/repo.go
