+++ options/locale/locale_en-US.ini
@@ -1716,7 +1716,7 @@ mirror_sync_create = synced new reference <a href="%s/src/%s">%[2]s</a> to <a hr
 mirror_sync_delete = synced and deleted reference <code>%[2]s</code> at <a href="%[1]s">%[3]s</a> from mirror
 migration_started = started a migration from %s to %s
 migration_successful = successfully migrated from %s to <a href="%s">%s</a>
-migration_failure = unsuccessfully migrated to %s: %s
+migration_failure = unsuccessfully migrated from %s to <a href="%s">%s</a>. You should delete this repository.
 
 [tool]
 ago = %s ago
