+++ templates/user/dashboard/feeds.tmpl
@@ -62,7 +62,7 @@
 						{{else if eq .GetOpType 22}}
 							{{$.i18n.Tr "action.migration_successful" .GetContent .GetRepoLink .ShortRepoPath | Str2html}}
 						{{else if eq .GetOpType 23}}
-							{{$.i18n.Tr "action.migration_failure" .ShortRepoPath .GetContent | Str2html}}
+							{{$.i18n.Tr "action.migration_failure" .GetContent .GetRepoLink .ShortRepoPath  | Str2html}}
 						{{end}}
 					</p>
 					{{if or (eq .GetOpType 5) (eq .GetOpType 18)}}
